Make "vcf_converter" file
